+++ src/main/java/org/jolokia/docker/maven/access/log/LogRequestor.java
@@ -68,8 +68,7 @@ public class LogRequestor extends Thread implements LogGetHandle {
         this.containerId = containerId;
         
         this.urlBuilder = urlBuilder;
-        
-        
+
         this.callback = callback;
         this.exception = null;
         this.setDaemon(true);