+++ doc/changelog.md
@@ -4,6 +4,7 @@
   - Allow CMD and ENTRYPOINT with shell and exec arguments (#130, #149)
   - Unix Socket support (#179)
   - Add a new parameter 'skipTags' for avoiding configured tagging of images (#145)
+  - Break build if log check or URL check runs into a timeout (#173)
   
 Please note that for consistencies sake `<command>` has been renamed to `<cmd>` which contains inner elements 
 to match better the equivalent Dockerfile argument. The update should be trivial and easy to spot since a build will croak immediately.